+++ app/Http/Livewire/Admin/User/Index.php
@@ -145,11 +145,14 @@ class Index extends Component
     {
         $deleteId = $event['data']['inputAttributes']['deleteId'];
         $model    = User::find($deleteId);
-        $model->delete();
-
-        $this->emit('refreshTable');
-
-        $this->alert('success', trans('messages.delete_success_message'));
+        if(!$model){
+            $this->emit('refreshTable'); 
+            $this->alert('error', trans('messages.error_message'));   
+        }else{
+            $model->delete();
+            $this->emit('refreshTable');    
+            $this->alert('success', trans('messages.delete_success_message'));
+        }     
     }
 
     public function toggle($id)
